backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
ultralytics
opencv-python
torch
torchvision
shapely
easyocr
python-multipart
pandas
matplotlib
torchreid
gdown
tensorboard
yacs
yt-dlp
redis
paddlepaddle
paddleocr
pydantic-settings
python-dotenv
httpx
numpy<2.0.0
twilio
requests
lap
